feat: change Format and FrameAncestorSettings references to Embed Settings
