+++ app/client/src/ce/utils/planHelpers.ts
@@ -1,6 +1,6 @@
 import type { FeatureFlags } from "@appsmith/entities/FeatureFlag";
 
-//if feature flag is true then return feature is enbaled
+//if feature flag is true then return feature is enabled
 export const isBrandingEnabled = (featureFlags: FeatureFlags) => {
   return featureFlags?.license_branding_enabled;
 };
