+++ app/client/src/pages/Applications/EmbedSnippet/Constants/constants.ts
@@ -1,0 +1,5 @@
+export enum AppsmithFrameAncestorsSetting {
+  ALLOW_EMBEDDING_EVERYWHERE = "ALLOW_EMBEDDING_EVERYWHERE",
+  LIMIT_EMBEDDING = "LIMIT_EMBEDDING",
+  DISABLE_EMBEDDING_EVERYWHERE = "DISABLE_EMBEDDING_EVERYWHERE",
+}
